+++ test_videos.py
@@ -171,17 +171,20 @@ def test(fp):
         # {"Blur/Sharpness": -60.46511627906977, "Contrast": 68.9186046511628, "Brightness": -6.976744186046517, "Smoothness": 96.54651162790698, "Binarize": true}
         # {"Blur/Sharpness": -46.162790697674424, "Contrast": 43.68604651162791, "Brightness": -46.51162790697675, "Smoothness": 38.2906976744186, "Binarize": true}
         # {"Blur/Sharpness": -46.162790697674424, "Contrast": 43.68604651162791, "Brightness": -46.51162790697675, "Smoothness": 38.2906976744186, "Binarize": False} - not bad with denoising
+        # {"Blur/Sharpness": -100.0, "Contrast": 36.68604651162791, "Brightness": -11.627906976744185, "Smoothness": 42.44186046511628, "Binarize": true, "Denoise SP": true}
+        # {"Blur/Sharpness": -100.0, "Contrast": 58.55813953488372, "Brightness": -4.6511627906976685, "Smoothness": 60.860465116279066, "Binarize": false, "Denoise SP": false}
+        # {"Blur/Sharpness": -100.0, "Contrast": 15.965116279069766, "Brightness": 0, "Smoothness": 30.930232558139533, "Binarize": false, "Denoise SP": true}
+        salt_pep_frame = tracking.preprocess_frame( scaled_frame, {"Blur/Sharpness": -100.0, "Contrast": 15.965116279069766, "Brightness": 0, "Smoothness": 30.930232558139533, "Binarize": False, "Denoise SP": True} , True)
 
-        salt_pep_frame = tracking.preprocess_frame( scaled_frame, {"Blur/Sharpness": -11.627906976744185, "Contrast": 49.348837209302324, "Brightness": -16.279069767441854, "Smoothness": 72.37209302325581, "Binarize": False} , True)
+        # saltpep_denoised = tracking.denoise_frame_saltpep(scaled_frame)
 
-        saltpep_denoised = tracking.denoise_frame_saltpep(salt_pep_frame)
-
+        thresholded = cv2.adaptiveThreshold(salt_pep_frame, 255, cv2.ADAPTIVE_THRESH_GAUSSIAN_C, cv2.THRESH_BINARY_INV, 11, 2)
 
         #frame_hstack_top = np.hstack((original, smoothedifuckinhope, binarizedifuckinhope))
         #frame_hstack_bottom = np.hstack((hp_filter_kernel, nlm_hp_filter_subtract, nlm_hp_filter_kernel))
         #frame_stack = np.vstack((frame_hstack_top, frame_hstack_bottom))
-        # frame_hstack = np.hstack((original, salt_pep_frame))
-        frame_hstack = np.hstack((original, saltpep_denoised))
+        frame_hstack = np.hstack((original, thresholded))
+        #frame_hstack = np.hstack((original, saltpep_denoised))
 
         cv2.imshow("Tracking...",frame_hstack)  # show updated frame tracking
 
@@ -196,6 +199,6 @@ def test(fp):
 
 if __name__ == '__main__':
     video_path = r"C:\Users\ipsou\ProgrammingStuff\Github\Marker-Tracker\viscoelastic_video\20240308_20240307_A549 _stiff VE sub._col-I_AS_current_01.vsi - 001 PH-Cell 1.avi"
-    # video_path = r"C:\Users\ipsou\ProgrammingStuff\Github\Marker-Tracker\viscoelastic_video\20240308_20240307_A549 _stiff VE sub._col-I_AS_current_02.vsi - 002 PH-Cell 2.avi"
+    #video_path = r"C:\Users\ipsou\ProgrammingStuff\Github\Marker-Tracker\viscoelastic_video\20240308_20240307_A549 _stiff VE sub._col-I_AS_current_02.vsi - 002 PH-Cell 2.avi"
     #video_path = r"C:\Users\ipsou\ProgrammingStuff\Github\Marker-Tracker\data\20240208_2024_07_02_A549.p23_PAHstiff_no beads_migration_exp.3_current_03.vsi - 003 PH.avi"
     test(video_path)